重构utils
