+++ src/MCDRpost/mcdrpost/manager/config_manager.py
@@ -4,8 +4,7 @@ from mcdreforged import PluginServerInterface
 
 from mcdrpost import constants
 from mcdrpost.configuration import Configuration
-from mcdrpost.utils import tr
-from mcdrpost.utils.translation_tags import Tags
+from mcdrpost.utils.translation import Tags, tr
 
 if TYPE_CHECKING:
     from mcdrpost.manager.post_manager import PostManager  # noqa
